+++ CHANGELOG.md
@@ -15,6 +15,7 @@
   - Fixed a bug that could lead to segmentation fault when a delete failed (#62)
   - Activity and Employment objects are now included in load log file (#47)
   - The rebuild-cache function now looks at "id" instead of "externalId" (#45)
+  - Better handling of when a CREATE fails because the object already existed (#44)
 
 ### v2.1.0 (2019-07-02)
 #### Bugfixes